ENH: implement features requested by experimentalists (#11)
* ENH: large fixes (see below)
* Logging, both to file and to screen. Enables debugging
* Dashboard, with some basic relevant plots.
* MAINT: add to makefile, add github CSS for tables

* ENH: support asking maximum number of queries

* ENH: allow uploading images or videos

* MAINT: correctly format ZIP targets

* BUG: correctly record timing information

* MAINT: get ready for deployment

* Add docs back in

* MAINT: deploy changes
* BUG: responses not recorded
* filename shown, warning presented if re-initializing
  and some javascript work

* Final AMI pushes

* MAINT: Disable buttons after click, add hover highlighting

+++ ami/salmon.sh
@@ -1,40 +1,50 @@
-# Get salmon
-sudo rm -rf /home/ubuntu/salmon
-git clone https://github.com/stsievert/salmon.git /home/ubuntu/salmon
-cd /home/ubuntu/salmon
+if [ -d "/home/ubuntu/salmon" ]; then
+    # Control will enter here if $DIRECTORY exists.
+    # This is the case when machine is launching for first time
+    #
+    # Install a working copy of Salmon. This copy on this machine will not be updated
+    # (if they want an update, they can start a new machine)
+    #
+    git clone https://github.com/stsievert/salmon.git /home/ubuntu/salmon
+    cd /home/ubuntu/salmon
+    git fetch --tags # Get new tags from remote
+    latestTag=$(git describe --tags `git rev-list --tags --max-count=1`) # Get latest tag name
+    git checkout $latestTag # Checkout latest tag
 
-# Get latest tag
-git fetch --tags # Get new tags from remote
-latestTag=$(git describe --tags `git rev-list --tags --max-count=1`) # Get latest tag name
-git checkout $latestTag # Checkout latest tag
-
-# Install fresh verison of Docker
-# https://docs.docker.com/install/linux/docker-ce/ubuntu/
-sudo apt-get remove -y docker docker-engine docker.io containerd runc
-sudo apt-get update
-sudo apt-get install -y apt-transport-https ca-certificates curl gnupg-agent software-properties-common
-curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
-sudo add-apt-repository \
-  "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
-  $(lsb_release -cs) \
-  stable"
-sudo apt-get update
-sudo apt-get install -y docker-ce docker-ce-cli containerd.io
-
-# install docker compose
-# https://docs.docker.com/compose/install/
-sudo curl -L "https://github.com/docker/compose/releases/download/1.24.1/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
-sudo chmod +x /usr/local/bin/docker-compose
+    # Make sure the latest salmon.sh is used
+    cd /home/ubuntu/salmon/ami
+    sudo mv salmon.service /lib/systemd/system/
+    sudo chmod u+x /home/ubuntu/salmon.sh
+    sudo systemctl enable salmon
+    cd /home/ubuntu/salmon
 
 
-cd /home/ubuntu/salmon;
-sudo docker-compose stop
-sudo docker-compose build
-sudo docker-compose up
+    # Install fresh verison of Docker
+    # https://docs.docker.com/install/linux/docker-ce/ubuntu/
+    sudo apt-get remove -y docker docker-engine docker.io containerd runc
+    sudo apt-get update
+    sudo apt-get install -y apt-transport-https ca-certificates curl gnupg-agent software-properties-common
+    curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
+    sudo add-apt-repository \
+      "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
+      $(lsb_release -cs) \
+      stable"
+    sudo apt-get update
+    sudo apt-get install -y docker-ce docker-ce-cli containerd.io
+
+    # install docker compose
+    # https://docs.docker.com/compose/install/
+    sudo curl -L "https://github.com/docker/compose/releases/download/1.24.1/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
+    sudo chmod +x /usr/local/bin/docker-compose
+
+    cd /home/ubuntu/salmon; sudo docker-compose build
+fi
+
+cd /home/ubuntu/salmon; sudo docker-compose up
 
 # Instructions for deploying to EC2: hhttps://askubuntu.com/questions/919054/how-do-i-run-a-single-command-at-startup-using-systemd
 # sudo mv salmon.service /lib/systemd/system/
 # sudo chmod u+x salmon.sh
 # sudo systemctl enable salmon
 #
-## View logs with `systemctl -l status salmon`
+## View logs with `systemctl -l status salmon
